requirements.txt
discord.py[voice]
gTTS
ffmpeg-python
requests
python-dotenv
